pyproject.toml
[project]
name = "starter-app-sample"
version = "0.2.0"
description = ""
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "langchain-google-vertexai>=2.0.7",
    "langchain==0.3.14",
    "langgraph==0.2.63",
    "langchain-core>=0.3.9",
    "langchain-google-community[vertexaisearch]>=2.0.2",
    "traceloop-sdk>=0.33.12",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "opentelemetry-sdk>=1.25.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-aiplatform[evaluation,langchain,reasoningengine]>=1.77.0",
    "scikit-learn>=1.5.0",
    "immutabledict>=4.2.0",
    "pypdf>=4.3.1",
    "langchain-openai>=0.2.10",
    "tomli>=2.2.1",
    "orjson==3.10.14",
]
requires-python = ">=3.9,<3.12"

[project.optional-dependencies]
streamlit = [
    "streamlit>=1.35.0",
    "streamlit-extras>=0.4.3",
    "extra-streamlit-components>=0.1.71",
    "streamlit-feedback>=0.1.3",
]
jupyter = [
    "jupyter>=1.0.0",
]
lint = [
    "flake8>=7.1.1",
    "flake8-pyproject>=1.2.3", 
    "mypy>=1",
    "codespell>=2.2.0",
    "black>=24.8.0",
    "pylint>=3.3.1",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,poetry.lock"

[tool.flake8]
ignore = ["E501", "W503"]

[tool.pylint]
disable = [
    "E0401",
    "C0301",
    "R0903",
    "R1710",
    "C0114",
    "R0915",
    "W1514",
    "W1203",
    "C0411",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
    "nest-asyncio>=1.6.0",
]

[tool.pytest.ini_options]
pythonpath = "."

[tool.hatch.build.targets.wheel]
packages = ["app","frontend"]
